refactor: S3 설정을 archive-repository-using-s3 모듈로 이동
- Worker에서 S3 설정을 재사용할 수 있도록 모듈을 이동하였습니다.
