+++ archive/archive-repository-using-s3/src/main/java/com/orv/archive/repository/s3/config/S3Config.java
@@ -1,37 +1,31 @@
-package com.orv.app.config;
+package com.orv.archive.repository.s3.config;
 
-import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
 
 @Configuration
-@Profile({"staging"}) // staging 환경에서만 로드됩니다.
-public class LocalStackS3Config {
-
+public class S3Config {
     @Value("${cloud.aws.credentials.accessKey}")
     private String accessKey;
-
     @Value("${cloud.aws.credentials.secretKey}")
     private String secretKey;
-
-    // LocalStack 사용시 region을 동일하게 설정하거나, 필요에 따라 변경
     @Value("${cloud.aws.region.static}")
     private String region;
 
     @Bean
+    @ConditionalOnMissingBean(AmazonS3.class)
     public AmazonS3 amazonS3Client() {
         BasicAWSCredentials awsCredentials = new BasicAWSCredentials(accessKey, secretKey);
         return AmazonS3ClientBuilder.standard()
-                .withEndpointConfiguration(
-                        new AwsClientBuilder.EndpointConfiguration("http://localstack:4566", region))
+                .withRegion(region)
                 .withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
-                .withPathStyleAccessEnabled(true) // LocalStack은 path-style access를 사용합니다.
+                .enableDualstack()
                 .build();
     }
 }
